Streamlit/pyproject.toml
[tool.poetry]
name = "crocodile-detector"
version = "0.1.0"
description = "Detection Crocodiles with YOLO V8"
authors = ["Lesego Matojane <lmatojane@gmail.com>"]
license = "MIT 2.0"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
numpy = "1.24.2"
opencv-python = "4.7.0.72"
pandas = "2.0.1"
pillow = "9.5.0"
streamlit = "1.22.0"
torch = "1.11.0"
ultralytics = "8.0.91"
click = "^8.1.3"


[tool.poetry.scripts]
crocodile-detector = 'cli:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
